pyproject.toml
[project]
name = "vanna2-analytics-tool"
version = "0.1.0"
description = "A multi-user, multi-turn SQL data analysis tool inspired by Vanna 2.0 (DeepSeek-chat + Qwen3-Embedding-8B)."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.7",
  "pydantic-settings>=2.2",
  "python-dotenv>=1.0",
  "httpx>=0.27",
  "SQLAlchemy[asyncio]>=2.0",
  "aiomysql>=0.2",
  "aiosqlite>=0.20",
  "python-jose[cryptography]>=3.3",
  "passlib[bcrypt]>=1.7",
  "chromadb>=0.5",
  "orjson>=3.10",
]

[project.optional-dependencies]
vanna = [
  "vanna>=2.0.0",
]

[dependency-groups]
dev = [
  "ruff>=0.6",
  "pytest>=8.0",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]
